Don't send the CSRF cookie to other pages
Add a path attribute to the CSRF cookie so that browsers only send it to
the redirect_uri page and not to other CKAN pages.

Potentially increases cacheability of pages, and also we may as well not
send our cookie with sensitive data in it unnecessarily.

This is difficult to add a test for so i haven't bothered.